implemented OrderService

+++ src/main/java/cs544/ea/OnlineRetailSystem/domain/CreditCard.java
@@ -2,7 +2,11 @@ package cs544.ea.OnlineRetailSystem.domain;
 
 import java.time.LocalDate;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.Data;
 
 @Data
@@ -13,7 +17,7 @@ public class CreditCard {
 	@Id
 	@GeneratedValue
 	@Column(name = "creditCardId")
-	private int id;
+	private Long id;
 	
 	@Column(name = "cardNumber")
 	private String cardNumber;
